Move the ready endpoint to its own module
Verbatim move of the /ready cache, probe, and mount out of the HTTP
transport module ahead of the per-request rewrite.
